requirements.txt
django==1.11.1
django-mptt==0.8.7
django-mptt-admin==0.4.5